requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3
drf-spectacular==0.26.4
Pillow==10.1.0
celery==5.3.4
redis==4.5.4
django-storages==1.14.1